MOVE: tests to test_assert_produces_warning
